Add sopc-parameter commandline option
used to automagically add parameters from sopc to the dts. choices are:
none: which does, well... nothing.
cmacro: add all "embeddedsw.CMacro.*" parameters
all: dump all assignments and parameters (this is insane and should not
be used other then being educated on the fact that you should not use it)

Signed-off-by: Walter Goossens <waltergoossens@home.nl>