sopc2dts: introduce boardinfo for offchip components
This is a fairly large patch that introduces the notion of boardinfo.
What this means is you get a small xml file describing all the stuff
about your board that can not be done using sopcbuilder and would
create huge commandlines to if we'd create parameters for these.
The usage is optional but it is recommended to take a look at the
example file (boardinfo_neek.xml) to see if this might be helpfull.
At this point you can define:
- Flash-layout for multiple chips (or a backup layout that is used when
no chip-name is supplied).
- Main memory. If you specify your main sd/ddr{1,2,3} ram and your tlb
memory here then all other memory components (onchip ssram bulkdata ddr
etc ) will be included in the dts like all other components.
- Kernel commandline
What will be added:
- i2c-slaves
- spi-slaves
- weird gpio assignments(?)
- more stuff you might want.
Everything is optional (as are most commandline options) and they are
handled int the order: commandline options, boardinfo settings, defaults.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>