sopc2dts: Create self-contained jar
This patch packs the sopc_components_* files with the jar and loads them
from the jar when running as a jar. When you run the classes without jar
the current directory is scanned for sopc_components_*.xml files to load.
If you changed the xml-files or want to add a custom components file, all
you have to do is rerun make and the jar will be updated.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>