sopc2dts: Add very basic PTF support
This patch adds some support for basic PTF parsing. This is mainly for nommu people that
want to "feel" some device-tree stuff without contacting the hw-department. This is not
yet very complete but more a proof of concept. PTF is missing quite a lot of info that
is present in sopcinfo files so upgrading is a better option.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>