sopc2dts: boardinfo: Add support for I2C
This patch enables you to describe the I2C chips connected to an i2c-master.
It basically works in the same way as creating flash-partitions, you either
name a specific master component or you can leave the name out and create a
wildcard description that matches all controllers. See boardinfo_neek.xml
for an example.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>