sopc2dts: Support hierarchical qsys designs
This patch enhances the support for loading sopcinfo files created
from a hierarchical qsys system. When a qsys subcomponent is found,
the directory containing the sopcinfo file is searched for the relevant
qsys file which is then loaded to get the finer details. It's not yet
complete but this is a good step in the right direction.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>