dts: Introduce new generator
This patch adds a new DTS/DTB generator. The old generator would produces text
directly that could then be compiled by an external devicetreecompiler. Because
some people use operating systems where it might be less trivial to compile the
devicetree compiler needed to generate the dtb, the generation of the dtb is
now integrated in sopc2dts using java.
The way to achieve this is to generate a list of devicetree nodes / properties
first from the loaded system and then transform these into a dts or dtb. The
changes required to get this to work are quite extensive, so chances are this
commit introduces some new bugs. It is tested against 25+ sopcinfo files that
were collected from the community and on first glance they work, but they're
not verified in hardware. If you experience any problems you can revert to the
old generators by specifying dts-old or dtb-old as the type.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>