sopc2dts: introduce TransparentInterfaceBridge
This patch adds the option to declare two (streaming for now) interfaces
and declare them as being transparently connected. This is usefull for fifo's
and monitoring/measuring cores that "listen" to a stream. Take a look at the
declaration of sc_fifo in sopc_components_altera.xml to see how it works.
Currently only the dts output works and only streaming interfaces are supported.
But it is a start...

Signed-off-by: Walter Goossens <waltergoossens@home.nl>